fix precision problem on orders

+++ coinex/coinex-client/src/main/scala/com/coinport/coinex/api/model/Order.scala
@@ -53,7 +53,7 @@ case class UserOrder(
         // regard total as quantity
         val quantity: Long = total match {
           case Some(t) => t.internalValue(currency2)
-          case None => (amount.get * price.get).internalValue(currency2)
+          case None => (BigDecimal(amount.get) * BigDecimal(price.get)).doubleValue().internalValue(currency2)
         }
         val limit = amount.map(_.internalValue(currency1))
         DoSubmitOrder(marketSide, Order(uid.toLong, id.toLong, quantity, newPrice, limit))
@@ -64,8 +64,8 @@ case class UserOrder(
           case Some(a) => a.internalValue(currency1)
           case None =>
             if (total.isDefined && price.isDefined) {
-              val totalValue = total.get / price.get
-              totalValue.internalValue(currency1)
+              val totalValue = BigDecimal(total.get) / BigDecimal(price.get)
+              totalValue.doubleValue().internalValue(currency1)
             } else 0
         }
         val limit = total map {
